Add fieldhandler base class
