+++ app/forms/handlers/select_multiple_handler.py
@@ -1,8 +1,8 @@
 from app.forms.custom_fields import CustomSelectMultipleField
-from app.forms.handlers.string_handler import StringHandler
+from app.forms.handlers.field_handler import FieldHandler
 
 
-class SelectMultipleHandler(StringHandler):
+class SelectMultipleHandler(FieldHandler):
     MANDATORY_MESSAGE = 'MANDATORY_CHECKBOX'
 
     def get_field(self):
